pyproject.toml
[project]
name = "langgraph-fabric-data-agent"
version = "0.1.0"
description = "LangGraph sample agent that calls a Fabric Data Agent over MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "azure-identity==1.24.0",
  "fastapi==0.116.1",
  "httpx==0.28.1",
  "langchain-core==0.3.78",
  "langchain-openai==0.3.35",
  "langgraph==0.6.8",
  "mcp==1.13.1",
  "microsoft-agents-activity==0.8.0",
  "microsoft-agents-authentication-msal==0.8.0",
  "microsoft-agents-hosting-aiohttp==0.8.0",
  "microsoft-agents-hosting-core==0.8.0",
  "pydantic==2.11.9",
  "pydantic-settings==2.11.0",
  "python-dotenv==1.1.1",
  "typer==0.17.4",
  "uvicorn==0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "pytest-asyncio==1.2.0",
  "pytest-cov==7.0.0",
  "respx==0.22.0",
  "ruff==0.13.2",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "N", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"
